refactor(scheduler): rename resourceaware plugin to gpujoborder for clarity
